requirements/base.txt
# API
fastapi==0.110.1
pydantic[email]==2.6.4
pydantic-settings==2.2.1
uvicorn==0.29.0
pyjwt[crypto]
bcrypt
python-multipart

# Database
alembic===1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.29

# Env
python-dotenv==1.0.1
